--- a/.travis.yml
+++ b/.travis.yml
@@ -45,6 +45,12 @@ matrix:
       - rustup component add clippy-preview
     script:
       - cargo clippy --all-targets --all-features -- -D warnings
+  - name: Benchmark
+    rust: stable
+    ? env
+    ? before_install
+    script:
+      - cargo bench
   - rust: stable
     env:
     before_install: