--- a/Cargo.toml
+++ b/Cargo.toml
@@ -39,3 +39,8 @@ rand = "0.3"
 rustc-test = "0.3.0"
 serde_json = "1.0"
 web3 = "0.4.0"
+criterion = "0.2"
+
+[[bench]]
+name = "transaction_bench"
+harness = false