As @jkilpatr noticed there is huge performance impact when sending transactions often. So this simple code does some benchmarking and with this simple change the performance has greatly improved.

Turns out we should reuse Secp256k1 context everywhere in clarity.

Before reusing context:

```
        sign tx without network id
                                time:   [14.749 ms 14.999 ms 15.247 ms]
        Found 7 outliers among 100 measurements (7.00%)
          6 (6.00%) high mild
          1 (1.00%) high severe
```

After reusing context:

```
        sign tx without network id
                                time:   [84.839 us 85.273 us 85.801 us]
                                change: [-99.412% -99.403% -99.393%] (p = 0.00 < 0.05)
                                Performance has improved.
    
```

99.4% improvement

Closes #20